backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.3.2

# AI/ML
openai==1.10.0
dashscope==1.14.0

# 工具库
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# 日志
python-json-logger==2.0.7
